# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Svet API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/calendar/depression": {
            "post": {
                "tags": [
                    "Календарь"
                ],
                "description": "Установка результатов теста на депрессию",
                "operationId": "74036eda504ff0ba9dd321ef150393af",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "json",
                        "in": "query",
                        "description": "JSON вида: {\"22.10.2022\":{\"depression\" : 90}, \"23.10.2022\":{\"depression\" : 30}}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/calendar/mood": {
            "post": {
                "tags": [
                    "Календарь"
                ],
                "description": "Установка настроений",
                "operationId": "5e8abc5bdc52d616e09070e648a572ef",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "json",
                        "in": "query",
                        "description": "JSON вида: {\"16.10.2022\" : {\"mood\" : [2,1,-1,-2]}, \"18.10.2022\" : {\"mood\" : [0,0,2,-2]}}\". Где числа в квадратных скобках означают отрезок настроения по цифровой шкале от -2 до 2 включая 0, где -2 это самая низкая оценка, а 2 самая высокая"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/notes/create": {
            "post": {
                "tags": [
                    "Заметки"
                ],
                "description": "Добавление заметки",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "note",
                        "in": "query",
                        "description": "Текст заметки"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/notes/delete": {
            "post": {
                "tags": [
                    "Заметки"
                ],
                "description": "Удаление заметки",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "note_id",
                        "in": "query",
                        "description": "ID заметки"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/trigger/create": {
            "post": {
                "tags": [
                    "Триггеры"
                ],
                "description": "Создание триггера",
                "operationId": "7749cc84cb886ba1fe4e893402a2508a",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "event",
                        "in": "query",
                        "description": "Событие"
                    },
                    {
                        "name": "feels",
                        "in": "query",
                        "description": "Ощущения"
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "description": "Действией"
                    },
                    {
                        "name": "thoughts",
                        "in": "query",
                        "description": "Мысли"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/trigger/thinking": {
            "post": {
                "tags": [
                    "Триггеры"
                ],
                "description": "Установка позитивного мышления спустя 24ч",
                "operationId": "d8743fbf81ab4234e2062cc48da6ab11",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "thinking",
                        "in": "query",
                        "description": "Описание"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/trigger/delete": {
            "post": {
                "tags": [
                    "Триггеры"
                ],
                "description": "Удаление триггера",
                "operationId": "8a09db92e54f94c77238e925034247c2",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "Защитный API токен"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID триггера"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
...
    }
}